新增了夜间模式

+++ app/src/main/java/skkk/gogogo/com/dakaizhihu/activity/SettingActivity.java
@@ -1,10 +1,12 @@
 package skkk.gogogo.com.dakaizhihu.activity;
 
 import android.app.AlertDialog;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.app.AppCompatDelegate;
 import android.support.v7.widget.Toolbar;
 import android.text.format.Formatter;
 import android.view.View;
@@ -35,6 +37,9 @@ public class SettingActivity extends AppCompatActivity {
     private SettingItemView sivImageMode;
     private AlertDialog clearDailog;
     private HorizortalProgressbarWithProgress progress;
+    private SettingItemView sivNightMode;
+    private boolean isNight;
+    private boolean dataStore;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -51,7 +56,15 @@ public class SettingActivity extends AppCompatActivity {
     * @时间 2016/7/24 11:32
     */
     private void beforeStart() {
-        mPref=getSharedPreferences("config",MODE_PRIVATE);
+        mPref = getSharedPreferences("config", MODE_PRIVATE);
+        isNight = mPref.getBoolean("night", false);
+        if (isNight) {
+            //设置为夜间模式
+            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
+        } else {
+            //设置为非夜间模式
+            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
+        }
     }
 
 
@@ -73,12 +86,12 @@ public class SettingActivity extends AppCompatActivity {
         * @desc 设置无图模式
         * @时间 2016/7/24 12:57
         */
-        sivImageMode= (SettingItemView) findViewById(R.id.siv_setting_image_mode);
-        boolean autoUpdate=mPref.getBoolean("image_mode",false);
+        sivImageMode = (SettingItemView) findViewById(R.id.siv_setting_image_mode);
+        boolean autoUpdate = mPref.getBoolean("image_mode", false);
 
-        if(autoUpdate){
+        if (autoUpdate) {
             sivImageMode.setChecked(true);
-        }else{
+        } else {
             sivImageMode.setChecked(false);
         }
 
@@ -101,6 +114,54 @@ public class SettingActivity extends AppCompatActivity {
             }
         });
 
+
+
+        /*
+        * @desc 设置夜间模式模式
+        * @时间 2016/7/24 12:57
+        */
+        sivNightMode = (SettingItemView) findViewById(R.id.siv_setting_night_mode);
+        isNight = mPref.getBoolean("night", false);
+
+        if (isNight) {
+            sivNightMode.setChecked(true);
+        } else {
+            sivNightMode.setChecked(false);
+        }
+
+        sivNightMode.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                //判断当前勾选状态
+                if (sivNightMode.isChecked()) {
+                    //设置不勾选
+                    sivNightMode.setChecked(false);
+                    //更新SP
+                    //设置为非夜间模式
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
+                    mPref.edit().putBoolean("night", false).commit();
+                    Toast.makeText(SettingActivity.this, "已关闭夜间模式模式", Toast.LENGTH_SHORT).show();
+                    //重启homeActivity
+                    Intent intent=new Intent();
+                    intent.setClass(SettingActivity.this, HomeActivity.class);
+                    startActivity(intent);
+                    finish();
+                } else {
+                    sivNightMode.setChecked(true);
+                    //更新SP
+                    //设置为夜间模式
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
+                    mPref.edit().putBoolean("night", true).commit();
+                    Toast.makeText(SettingActivity.this, "已开启夜间模式", Toast.LENGTH_SHORT).show();
+                    //重启homeActivity
+                    Intent intent=new Intent();
+                    intent.setClass(SettingActivity.this,HomeActivity.class);
+                    startActivity(intent);
+                    finish();
+                }
+            }
+        });
+
     }
 
     private void initDB() {
@@ -108,60 +169,63 @@ public class SettingActivity extends AppCompatActivity {
     }
 
 
+    public void clearSQL(View view) {
+        dataStore = mPref.getBoolean("data_store",true);
+        if (dataStore){
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setTitle("提示");
+            builder.setMessage("正在清理缓存...");
+            View view_2 = View.inflate(this, R.layout.view_progress, null);
+            progress = (HorizortalProgressbarWithProgress) view_2.findViewById(R.id.my_progress);
 
+            //删除数据库数据
+            db = helper.getWritableDatabase();
+            db.delete("News", null, null);
+            builder.setView(view_2);
+            new Thread(new Runnable() {
+                @Override
+                public void run() {
+                    db.delete("News", null, null);
+                    for (int i = 1; i <= 100; i++) {
+                        try {
+                            Thread.sleep(50);
+                        } catch (InterruptedException e) {
+                            e.printStackTrace();
+                        }
+                        progress.setProgress(i);//设置进度
+                    }
+                    clearDailog.dismiss();
+                    runOnUiThread(new Runnable() {
+                        @Override
+                        public void run() {
+                            tvSettingSize.setText("0K");
+                            mPref.edit().putBoolean("data_store",false).commit();
+                        }
+                    });
+                }
+            }).start();
+            clearDailog = builder.show();
+        }else{
+            Toast.makeText(SettingActivity.this, "暂无缓存", Toast.LENGTH_SHORT).show();
+        }
 
-
-    public void clearSQL(View view){
-
-        AlertDialog.Builder builder=new AlertDialog.Builder(this);
-        builder.setTitle("提示");
-        builder.setMessage("正在清理缓存...");
-        View view_2=View.inflate(this,R.layout.view_progress,null);
-        progress = (HorizortalProgressbarWithProgress) view_2.findViewById(R.id.my_progress);
-
-        //删除数据库数据
-        db=helper.getWritableDatabase();
-        db.delete("News", null, null);
-        builder.setView(view_2);
-        new Thread(new Runnable() {
-             @Override
-             public void run() {
-                 db.delete("News", null, null);
-                 for (int i=1;i<=100;i++){
-                     try {
-                         Thread.sleep(50);
-                     } catch (InterruptedException e) {
-                         e.printStackTrace();
-                     }
-                     progress.setProgress(i);//设置进度
-                 }
-                 clearDailog.dismiss();
-                 runOnUiThread(new Runnable() {
-                     @Override
-                     public void run() {
-                         tvSettingSize.setText("0K");
-                     }
-                 });
-             }
-         }).start();
-        clearDailog=builder.show();
 
     }
 
     private void initData() {
 
         try {
-            File file=new File("/data/data/"
+            File file = new File("/data/data/"
                     + getPackageName() + "/databases");
             l = getFileSize.getFolderSize(file);
 
         } catch (Exception e) {
             e.printStackTrace();
         }
-        if(l<50*1024){
+        if (l < 50 * 1024) {
             tvSettingSize.setText("0K");
-        }else{
-            String fileSize= Formatter.formatFileSize(this, l);
+        } else {
+            String fileSize = Formatter.formatFileSize(this, l);
             tvSettingSize.setText(fileSize);
         }
